+++ CHANGES.md
@@ -1,5 +1,7 @@
 #### Version 1.1
 - Added batch downloader
+- Added option to create comment when bigger version of image was found
+- Added option to use stored source URL when updating tags on existing posts
 
 #### Version 1.0
 - Initial release